feat: Add functionality to create new articles
This commit adds the functionality to create new articles in the CreatePost page. It includes the necessary changes to the useFirebase hook and the CreatePost component. The user can now enter a title, description, and image URL for the new article and submit it. The article is then added to the "articles" collection in the Firebase Firestore database.