feat: Enable image upload in CreatePost page
This commit enables the functionality to upload an image in the CreatePost page. The user can now select an image file and it will be associated with the new article being created. The image URL is stored in the "imageUrl" field of the article object and is added to the Firestore database along with the title, description, and creation date.