feat: Add dropdown component to CreatePost page
The code changes introduce a new dropdown component to the CreatePost page. This dropdown component will provide additional options for the user when creating a new post. The dropdown component is imported from the "ui-elements/Dropdown" module and is used in the CreatePost component. This enhancement improves the user experience by allowing users to select additional options for their posts.